Move measurement components to separate module.
