+++ components/frontend/src/measurement/StatusIcon.css
