+++ components/frontend/src/measurement/StatusIcon.js
