+++ components/frontend/src/measurement/StatusIcon.test.js
