+++ components/frontend/src/measurement/TrendSparkline.js
