+++ components/frontend/src/measurement/TrendSparkline.test.js
