--- components/frontend/src/metric/Metric.css
@@ -1,23 +1,0 @@
-tr.target_met {
-    background-color: rgb(30, 148, 78, 0.15) !important;
-}
-
-tr.target_not_met {
-    background-color: rgb(211, 59, 55, 0.2) !important;
-}
-
-tr.near_target_met {
-    background-color: rgb(253, 197, 54, 0.15) !important;
-}
-
-tr.debt_target_met {
-    background-color: rgb(150, 150, 150, 0.2) !important;
-}
-
-tr.unknown {
-    background-color: rgb(245, 245, 245, 0.15) !important;
-}
-
-td > a {
-    color: rgb(0, 88, 176) !important;
-}
